Wire logic for removing book from display
e.path allows access to the path through which the event bubbles. In present
case we can access the relevant book div by looking at the first element
encountered upstream.

Note also, in principle, the book array is not needed to enable functionality
for just displaying and deleting added books.